+++ src/main/java/me/anedhel/lotr/block/ModBlocks.java
@@ -1,11 +1,13 @@
 package me.anedhel.lotr.block;
 
 import me.anedhel.lotr.LordOfTheRingsMiddleEarthMod;
+import me.anedhel.lotr.block.custom.crops.CornCropBlock;
 import me.anedhel.lotr.block.custom.crops.LettuceCropBlock;
 import me.anedhel.lotr.block.custom.crops.TomatoCropBlock;
 import net.fabricmc.fabric.api.item.v1.FabricItemSettings;
 import net.fabricmc.fabric.api.object.builder.v1.block.FabricBlockSettings;
 import net.minecraft.block.*;
+import net.minecraft.entity.effect.StatusEffects;
 import net.minecraft.item.BlockItem;
 import net.minecraft.item.Item;
 import net.minecraft.registry.Registries;
@@ -30,10 +32,46 @@ public class ModBlocks {
             new TomatoCropBlock(FabricBlockSettings.copyOf(Blocks.WHEAT)));
     public static final Block LETTUCE_CROP = Registry.register(Registries.BLOCK, new Identifier(LordOfTheRingsMiddleEarthMod.MOD_ID, "lettuce_crop"),
             new LettuceCropBlock(FabricBlockSettings.copyOf(Blocks.BEETROOTS)));
+    public static final Block CORN_CROP = Registry.register(Registries.BLOCK, new Identifier(LordOfTheRingsMiddleEarthMod.MOD_ID, "corn_crop"),
+            new CornCropBlock(FabricBlockSettings.copyOf(Blocks.WHEAT)));
+
+    public static final Block WILD_BEETROOT = registerBlock("wild_beetroot",
+            new FlowerBlock(StatusEffects.SATURATION, 10, FabricBlockSettings.copyOf(Blocks.ALLIUM).nonOpaque().noCollision()));
+    public static final Block POTTED_WILD_BEETROOT = Registry.register(Registries.BLOCK, new Identifier(LordOfTheRingsMiddleEarthMod.MOD_ID, "potted_wild_beetroot"),
+            new FlowerPotBlock(WILD_BEETROOT, FabricBlockSettings.copyOf(Blocks.POTTED_ALLIUM).nonOpaque()));
+    public static final Block WILD_CARROT = registerBlock("wild_carrot",
+            new FlowerBlock(StatusEffects.SATURATION, 10, FabricBlockSettings.copyOf(Blocks.ALLIUM).nonOpaque().noCollision()));
+    public static final Block POTTED_WILD_CARROT = Registry.register(Registries.BLOCK, new Identifier(LordOfTheRingsMiddleEarthMod.MOD_ID, "potted_wild_carrot"),
+            new FlowerPotBlock(WILD_CARROT, FabricBlockSettings.copyOf(Blocks.POTTED_ALLIUM).nonOpaque()));
+    public static final Block WILD_POTATO = registerBlock("wild_potato",
+            new FlowerBlock(StatusEffects.SATURATION, 10, FabricBlockSettings.copyOf(Blocks.ALLIUM).nonOpaque().noCollision()));
+    public static final Block POTTED_WILD_POTATO = Registry.register(Registries.BLOCK, new Identifier(LordOfTheRingsMiddleEarthMod.MOD_ID, "potted_wild_potato"),
+            new FlowerPotBlock(WILD_POTATO, FabricBlockSettings.copyOf(Blocks.POTTED_ALLIUM).nonOpaque()));
+    public static final Block WILD_CORN = registerBlock("wild_corn",
+            new FlowerBlock(StatusEffects.SATURATION, 10, FabricBlockSettings.copyOf(Blocks.ALLIUM).nonOpaque().noCollision()));
+    public static final Block POTTED_WILD_CORN = Registry.register(Registries.BLOCK, new Identifier(LordOfTheRingsMiddleEarthMod.MOD_ID, "potted_wild_corn"),
+            new FlowerPotBlock(WILD_CORN, FabricBlockSettings.copyOf(Blocks.POTTED_ALLIUM).nonOpaque()));
+    public static final Block WILD_LETTUCE = registerBlock("wild_lettuce",
+            new FlowerBlock(StatusEffects.SATURATION, 10, FabricBlockSettings.copyOf(Blocks.ALLIUM).nonOpaque().noCollision()));
+    public static final Block POTTED_WILD_LETTUCE = Registry.register(Registries.BLOCK, new Identifier(LordOfTheRingsMiddleEarthMod.MOD_ID, "potted_wild_lettuce"),
+            new FlowerPotBlock(WILD_LETTUCE, FabricBlockSettings.copyOf(Blocks.POTTED_ALLIUM).nonOpaque()));
+    public static final Block WILD_TOMATO = registerBlock("wild_tomato",
+            new FlowerBlock(StatusEffects.SATURATION, 10, FabricBlockSettings.copyOf(Blocks.ALLIUM).nonOpaque().noCollision()));
+    public static final Block POTTED_WILD_TOMATO = Registry.register(Registries.BLOCK, new Identifier(LordOfTheRingsMiddleEarthMod.MOD_ID, "potted_wild_tomato"),
+            new FlowerPotBlock(WILD_TOMATO, FabricBlockSettings.copyOf(Blocks.POTTED_ALLIUM).nonOpaque()));
+
+    public static final Block POTATO_CRATE = registerBlock("potato_crate",
+            new Block(FabricBlockSettings.copyOf(Blocks.COMPOSTER)));
+    public static final Block CARROT_CRATE = registerBlock("carrot_crate",
+            new Block(FabricBlockSettings.copyOf(Blocks.COMPOSTER)));
+    public static final Block BEETROOT_CRATE = registerBlock("beetroot_crate",
+            new Block(FabricBlockSettings.copyOf(Blocks.COMPOSTER)));
 
     public static final Block TOMATO_CRATE = registerBlock("tomato_crate",
             new Block(FabricBlockSettings.copyOf(Blocks.COMPOSTER)));
     public static final Block LETTUCE_CRATE = registerBlock("lettuce_crate",
+            new Block(FabricBlockSettings.copyOf(Blocks.COMPOSTER)));
+    public static final Block CORN_CRATE = registerBlock("corn_crate",
             new Block(FabricBlockSettings.copyOf(Blocks.COMPOSTER)));
 
     public static final Block CHALK = registerBlock("chalk",