+++ src/main/java/me/anedhel/lotr/datagen/ModRecipeProvider.java
@@ -8,7 +8,7 @@ import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
 import net.fabricmc.fabric.api.datagen.v1.provider.FabricRecipeProvider;
 import net.minecraft.block.Block;
 import net.minecraft.block.Blocks;
-import net.minecraft.data.server.recipe.RecipeJsonProvider;
+import net.minecraft.data.server.recipe.RecipeExporter;
 import net.minecraft.data.server.recipe.ShapedRecipeJsonBuilder;
 import net.minecraft.data.server.recipe.ShapelessRecipeJsonBuilder;
 import net.minecraft.item.Item;
@@ -18,7 +18,6 @@ import net.minecraft.recipe.book.RecipeCategory;
 import net.minecraft.util.Identifier;
 
 import java.util.List;
-import java.util.function.Consumer;
 
 /**
  * This class is used to register all crafting recipes the mod provides.
@@ -40,7 +39,7 @@ public class ModRecipeProvider extends FabricRecipeProvider {
      * @param exporter A RecipeJsonProvider, which is provided by minecraft
      */
     @Override
-    public void generate(Consumer<RecipeJsonProvider> exporter) {
+    public void generate(RecipeExporter exporter) {
         offerSmelting(exporter, TIN_SMELTABLES, RecipeCategory.MISC, ModItems.TIN_INGOT, 0.7f, 200, "tin");
         offerBlasting(exporter, TIN_SMELTABLES, RecipeCategory.MISC, ModItems.TIN_INGOT, 0.7f, 100, "tin");
         offerSmelting(exporter, TIN_BLOCK_SMELTABLES, RecipeCategory.MISC, ModBlocks.TIN_BLOCK, 5.6f, 1600, "tin_block");
@@ -64,8 +63,22 @@ public class ModRecipeProvider extends FabricRecipeProvider {
         offerReversibleCompactingRecipes(exporter, RecipeCategory.BUILDING_BLOCKS, ModItems.TIN_INGOT, RecipeCategory.MISC, ModBlocks.TIN_BLOCK);
         offerReversibleCompactingRecipes(exporter, RecipeCategory.BUILDING_BLOCKS, ModItems.BRONZE_INGOT, RecipeCategory.MISC, ModBlocks.BRONZE_BLOCK);
 
+        ShapelessRecipeJsonBuilder.create(RecipeCategory.MISC, ModItems.TOMATO_SEEDS, 1)
+                        .input(ModItems.TOMATO)
+                        .criterion(hasItem(ModItems.TOMATO), conditionsFromItem(ModItems.TOMATO))
+                        .offerTo(exporter, new Identifier(getRecipeName(ModItems.TOMATO_SEEDS)));
+
+        ShapelessRecipeJsonBuilder.create(RecipeCategory.MISC, ModItems.CORN_SEEDS, 1)
+                        .input(ModItems.CORN)
+                        .criterion(hasItem(ModItems.CORN), conditionsFromItem(ModItems.CORN))
+                        .offerTo(exporter, new Identifier(getRecipeName(ModItems.CORN_SEEDS)));
+
         createCrateRecipe(ModItems.TOMATO, ModBlocks.TOMATO_CRATE, exporter);
         createCrateRecipe(ModItems.LETTUCE, ModBlocks.LETTUCE_CRATE, exporter);
+        createCrateRecipe(ModItems.CORN, ModBlocks.CORN_CRATE, exporter);
+        createCrateRecipe(Items.POTATO, ModBlocks.POTATO_CRATE, exporter);
+        createCrateRecipe(Items.CARROT, ModBlocks.CARROT_CRATE, exporter);
+        createCrateRecipe(Items.BEETROOT, ModBlocks.BEETROOT_CRATE, exporter);
 
         createModGearTypeRecipes(ModGearType.BRONZE, exporter);
 
@@ -101,7 +114,7 @@ public class ModRecipeProvider extends FabricRecipeProvider {
      * @param gearType The ModGearType that all craftingRecipes are to be generated for
      * @param exporter The exporter is an instance you offer the crafting recipe to. Usually one is provided in the parameters of the method you edit.
      */
-    private void createModGearTypeRecipes (ModGearType gearType, Consumer<RecipeJsonProvider> exporter) {
+    private void createModGearTypeRecipes (ModGearType gearType, RecipeExporter exporter) {
         Item material = gearType.getMaterial();
         Item helmet = gearType.getHelmet();
         Item chestplate = gearType.getChestplate();
@@ -151,7 +164,7 @@ public class ModRecipeProvider extends FabricRecipeProvider {
      * @param cobbledVariant The cobbled version of the stoneType, null if there is none
      * @param brickVariant The bricks of the stoneType, null if there is none
      */
-    private void createModStoneTypeRecipes (ModStoneType stoneType, Consumer<RecipeJsonProvider> exporter, ModStoneType cobbledVariant, ModStoneType brickVariant) {
+    private void createModStoneTypeRecipes (ModStoneType stoneType, RecipeExporter exporter, ModStoneType cobbledVariant, ModStoneType brickVariant) {
         Block baseBlock = stoneType.getBaseBlock();
         Block stairBlock = stoneType.getStairBlock();
         Block slabBlock = stoneType.getSlabBlock();
@@ -199,7 +212,13 @@ public class ModRecipeProvider extends FabricRecipeProvider {
         }
     }
 
-    private void createCrateRecipe (Item item, Block crate, Consumer<RecipeJsonProvider> exporter) {
+    /**
+     * This Method is used to generate a Crafting Recipe for a StorageCrate of the FarmItem
+     * @param item The Farm Item you wish to compact into a crate
+     * @param crate The Crate for the Farm Item
+     * @param exporter The exporter is an instance you offer the crafting recipe to. Usually one is provided in the parameters of the method you edit.
+     */
+    private void createCrateRecipe (Item item, Block crate, RecipeExporter exporter) {
         ShapedRecipeJsonBuilder.create(RecipeCategory.DECORATIONS, crate, 1)
                 .pattern("###")
                 .pattern("###")
@@ -221,7 +240,7 @@ public class ModRecipeProvider extends FabricRecipeProvider {
      * @param output the brick of the blank stone type
      * @param exporter The exporter is an instance you offer the crafting recipe to. Usually one is provided in the parameters of the method you edit.
      */
-    private void createBricksRecipe (Block input, Block output, Consumer<RecipeJsonProvider> exporter) {
+    private void createBricksRecipe (Block input, Block output, RecipeExporter exporter) {
         ShapedRecipeJsonBuilder.create(RecipeCategory.BUILDING_BLOCKS, output, 4)
                 .pattern("##")
                 .pattern("##")
@@ -236,7 +255,7 @@ public class ModRecipeProvider extends FabricRecipeProvider {
      * @param output the cobbled variant of the blank stone type
      * @param exporter The exporter is an instance you offer the crafting recipe to. Usually one is provided in the parameters of the method you edit.
      */
-    private void createCobbledStoneTypeRecipe (Block input, Block output, Consumer<RecipeJsonProvider> exporter) {
+    private void createCobbledStoneTypeRecipe (Block input, Block output, RecipeExporter exporter) {
         ShapedRecipeJsonBuilder.create(RecipeCategory.BUILDING_BLOCKS, output, 4)
                 .pattern("#C")
                 .pattern("C#")
@@ -253,7 +272,7 @@ public class ModRecipeProvider extends FabricRecipeProvider {
      * @param output the button of the blank stone type
      * @param exporter The exporter is an instance you offer the crafting recipe to. Usually one is provided in the parameters of the method you edit.
      */
-    private void createButtonRecipe (Block input, Block output, Consumer<RecipeJsonProvider> exporter) {
+    private void createButtonRecipe (Block input, Block output, RecipeExporter exporter) {
         ShapedRecipeJsonBuilder.create(RecipeCategory.REDSTONE, output, 1)
                 .pattern(" # ")
                 .input('#', input)
@@ -267,7 +286,7 @@ public class ModRecipeProvider extends FabricRecipeProvider {
      * @param output the stairs of the input
      * @param exporter The exporter is an instance you offer the crafting recipe to. Usually one is provided in the parameters of the method you edit.
      */
-    private void createStairRecipe (Block input, Block output, Consumer<RecipeJsonProvider> exporter) {
+    private void createStairRecipe (Block input, Block output, RecipeExporter exporter) {
         ShapedRecipeJsonBuilder.create(RecipeCategory.BUILDING_BLOCKS, output, 4)
                 .pattern("#  ")
                 .pattern("## ")
@@ -283,7 +302,7 @@ public class ModRecipeProvider extends FabricRecipeProvider {
      * @param output The sword made from the Ingredient
      * @param exporter The exporter is an instance you offer the crafting recipe to. Usually one is provided in the parameters of the method you edit.
      */
-    private void createSwordRecipe (Item input, Item output, Consumer<RecipeJsonProvider> exporter){
+    private void createSwordRecipe (Item input, Item output, RecipeExporter exporter){
         ShapedRecipeJsonBuilder.create(RecipeCategory.BUILDING_BLOCKS, output, 1)
                 .pattern("#")
                 .pattern("#")
@@ -301,7 +320,7 @@ public class ModRecipeProvider extends FabricRecipeProvider {
      * @param output The axe made from the Ingredient
      * @param exporter The exporter is an instance you offer the crafting recipe to. Usually one is provided in the parameters of the method you edit.
      */
-    private void createAxeRecipe (Item input, Item output, Consumer<RecipeJsonProvider> exporter){
+    private void createAxeRecipe (Item input, Item output, RecipeExporter exporter){
         ShapedRecipeJsonBuilder.create(RecipeCategory.BUILDING_BLOCKS, output, 1)
                 .pattern("## ")
                 .pattern("#S ")
@@ -319,7 +338,7 @@ public class ModRecipeProvider extends FabricRecipeProvider {
      * @param output The pickaxe made from the Ingredient
      * @param exporter The exporter is an instance you offer the crafting recipe to. Usually one is provided in the parameters of the method you edit.
      */
-    private void createPickaxeRecipe (Item input, Item output, Consumer<RecipeJsonProvider> exporter){
+    private void createPickaxeRecipe (Item input, Item output, RecipeExporter exporter){
         ShapedRecipeJsonBuilder.create(RecipeCategory.BUILDING_BLOCKS, output, 1)
                 .pattern("###")
                 .pattern(" S ")
@@ -337,7 +356,7 @@ public class ModRecipeProvider extends FabricRecipeProvider {
      * @param output The shovel made from the Ingredient
      * @param exporter The exporter is an instance you offer the crafting recipe to. Usually one is provided in the parameters of the method you edit.
      */
-    private void createShovelRecipe (Item input, Item output, Consumer<RecipeJsonProvider> exporter){
+    private void createShovelRecipe (Item input, Item output, RecipeExporter exporter){
         ShapedRecipeJsonBuilder.create(RecipeCategory.BUILDING_BLOCKS, output, 1)
                 .pattern("#")
                 .pattern("S")
@@ -355,7 +374,7 @@ public class ModRecipeProvider extends FabricRecipeProvider {
      * @param output The how made from the Ingredient
      * @param exporter The exporter is an instance you offer the crafting recipe to. Usually one is provided in the parameters of the method you edit.
      */
-    private void createHoeRecipe (Item input, Item output, Consumer<RecipeJsonProvider> exporter){
+    private void createHoeRecipe (Item input, Item output, RecipeExporter exporter){
         ShapedRecipeJsonBuilder.create(RecipeCategory.BUILDING_BLOCKS, output, 1)
                 .pattern("## ")
                 .pattern(" S ")
@@ -373,7 +392,7 @@ public class ModRecipeProvider extends FabricRecipeProvider {
      * @param output The helmet made from the Ingredient
      * @param exporter The exporter is an instance you offer the crafting recipe to. Usually one is provided in the parameters of the method you edit.
      */
-    private void createHelmetRecipe(Item input, Item output, Consumer<RecipeJsonProvider> exporter) {
+    private void createHelmetRecipe(Item input, Item output, RecipeExporter exporter) {
         ShapedRecipeJsonBuilder.create(RecipeCategory.BUILDING_BLOCKS, output, 1)
                 .pattern("###")
                 .pattern("# #")
@@ -388,7 +407,7 @@ public class ModRecipeProvider extends FabricRecipeProvider {
      * @param output The chestplate made from the Ingredient
      * @param exporter The exporter is an instance you offer the crafting recipe to. Usually one is provided in the parameters of the method you edit.
      */
-    private void createChestplateRecipe(Item input, Item output, Consumer<RecipeJsonProvider> exporter) {
+    private void createChestplateRecipe(Item input, Item output, RecipeExporter exporter) {
         ShapedRecipeJsonBuilder.create(RecipeCategory.BUILDING_BLOCKS, output, 1)
                 .pattern("# #")
                 .pattern("###")
@@ -404,7 +423,7 @@ public class ModRecipeProvider extends FabricRecipeProvider {
      * @param output The leggins made from the Ingredient
      * @param exporter The exporter is an instance you offer the crafting recipe to. Usually one is provided in the parameters of the method you edit.
      */
-    private void createLegginsRecipe(Item input, Item output, Consumer<RecipeJsonProvider> exporter) {
+    private void createLegginsRecipe(Item input, Item output, RecipeExporter exporter) {
         ShapedRecipeJsonBuilder.create(RecipeCategory.BUILDING_BLOCKS, output, 1)
                 .pattern("###")
                 .pattern("# #")
@@ -420,7 +439,7 @@ public class ModRecipeProvider extends FabricRecipeProvider {
      * @param output The boots made from the Ingredient
      * @param exporter The exporter is an instance you offer the crafting recipe to. Usually one is provided in the parameters of the method you edit.
      */
-    private void createBootsRecipe(Item input, Item output, Consumer<RecipeJsonProvider> exporter) {
+    private void createBootsRecipe(Item input, Item output, RecipeExporter exporter) {
         ShapedRecipeJsonBuilder.create(RecipeCategory.BUILDING_BLOCKS, output, 1)
                 .pattern("# #")
                 .pattern("# #")