+++ src/main/java/me/anedhel/lotr/item/ModFoodComponents.java
@@ -5,5 +5,5 @@ import net.minecraft.item.FoodComponent;
 public class ModFoodComponents {
 
     public static final FoodComponent TOMATO = new FoodComponent.Builder().hunger(3).saturationModifier(0.25f).alwaysEdible().build();
-
+    public static final FoodComponent CORN = new FoodComponent.Builder().hunger(3).saturationModifier(0.25f).alwaysEdible().build();
 }