+++ CHANGELOG.md
@@ -18,3 +18,7 @@ ___
 + Added Smooth Basalt Building Blocks
 + Added Cobbled Blueslate with the building blocks
 + Added Tomato and Lettuce as Crops, with respective crates as Compacted Block
++ Added Corn as a 2 Block high crop with the respective crate as Compacted Block
++ Added Crates for Beetroot, Carrots and Potatoes
++ Added Wild Flowers for Beetroot, Carrots, Potatoes, Corn, Lettuce and Tomato, that drop the Respective Items or themselves when sheared
++ Added Corn as Composter usable Item